refactor: rename core to production 2
